qt: Rename configuration_web to "configuration_network"
