+++ src/citra_qt/configuration/configure_dialog.cpp
@@ -14,10 +14,10 @@
 #include "citra_qt/configuration/configure_hotkeys.h"
 #include "citra_qt/configuration/configure_input.h"
 #include "citra_qt/configuration/configure_layout.h"
+#include "citra_qt/configuration/configure_network.h"
 #include "citra_qt/configuration/configure_storage.h"
 #include "citra_qt/configuration/configure_system.h"
 #include "citra_qt/configuration/configure_ui.h"
-#include "citra_qt/configuration/configure_web.h"
 #include "citra_qt/hotkeys.h"
 #include "common/settings.h"
 #include "core/core.h"
